style(post): 게시글 도메인 이름, 게시글 등록 요청 경로 변경 (#2)
