+++ src/api/config.js
@@ -1,6 +1,6 @@
 export const API_SERVER_HOST = 'http://localhost:8080'
 
-export const BOARD = `${API_SERVER_HOST}/board`
+export const POST = `${API_SERVER_HOST}/posts`
 
 export const USER = `${API_SERVER_HOST}/users`
