+++ src/api/postApi.js
@@ -1,8 +1,8 @@
-import {BOARD} from "./config.js";
+import {POST} from "./config.js";
 import axios from "axios";
 
 export const postAdd = async (addParam) => {
     const header = {headers: {'Content-Type': 'application/json'}}
 
-    return await axios.post(`${BOARD}`, addParam, header)
+    return await axios.post(`${POST}`, addParam, header)
 }
