--- src/components/board/list/ListBasicComponent.jsx
@@ -1,122 +1,0 @@
-import React from 'react';
-import ImageListItem from "@mui/material/ImageListItem";
-import ListSubheader from "@mui/material/ListSubheader";
-import ImageListItemBar from "@mui/material/ImageListItemBar";
-import ShareMenu from "../../menus/share/ShareMenu.jsx";
-import ImageList from "@mui/material/ImageList";
-
-function ListBasicComponent() {
-
-
-  return (
-
-      <ImageList sx={{width: 1300, height: 650}}>
-        <ImageListItem key="Subheader" cols={2}>
-          <ListSubheader component="div">트렌딩 보드 </ListSubheader>
-        </ImageListItem>
-        {itemData.map((item) => (
-            <ImageListItem key={item.img}>
-              <img
-                  srcSet={`${item.img}?w=248&fit=crop&auto=format&dpr=2 2x`}
-                  src={`${item.img}?w=248&fit=crop&auto=format`}
-                  alt={item.title}
-                  loading="lazy"
-              />
-              <ImageListItemBar
-                  title={item.title}
-                  subtitle={item.author}
-                  actionIcon={<ShareMenu id={item.id}/>}
-              />
-            </ImageListItem>
-        ))}
-      </ImageList>
-  );
-}
-
-const itemData = [
-  {
-    id: 1,
-    img: 'https://images.unsplash.com/photo-1551963831-b3b1ca40c98e',
-    title: 'Breakfast',
-    author: '@bkristastucchio',
-    rows: 2,
-    cols: 2,
-    featured: true,
-  },
-  {
-    id: 2,
-    img: 'https://images.unsplash.com/photo-1551782450-a2132b4ba21d',
-    title: 'Burger',
-    author: '@rollelflex_graphy726',
-  },
-  {
-    id: 3,
-    img: 'https://images.unsplash.com/photo-1522770179533-24471fcdba45',
-    title: 'Camera',
-    author: '@helloimnik',
-  },
-  {
-    id: 4,
-    img: 'https://images.unsplash.com/photo-1444418776041-9c7e33cc5a9c',
-    title: 'Coffee',
-    author: '@nolanissac',
-    cols: 2,
-  },
-  {
-    id: 5,
-    img: 'https://images.unsplash.com/photo-1533827432537-70133748f5c8',
-    title: 'Hats',
-    author: '@hjrc33',
-    cols: 2,
-  },
-  {
-    id: 6,
-    img: 'https://images.unsplash.com/photo-1558642452-9d2a7deb7f62',
-    title: 'Honey',
-    author: '@arwinneil',
-    rows: 2,
-    cols: 2,
-    featured: true,
-  },
-  {
-    id: 7,
-    img: 'https://images.unsplash.com/photo-1516802273409-68526ee1bdd6',
-    title: 'Basketball',
-    author: '@tjdragotta',
-  },
-  {
-    id: 8,
-    img: 'https://images.unsplash.com/photo-1518756131217-31eb79b20e8f',
-    title: 'Fern',
-    author: '@katie_wasserman',
-  },
-  {
-    id: 9,
-    img: 'https://images.unsplash.com/photo-1597645587822-e99fa5d45d25',
-    title: 'Mushrooms',
-    author: '@silverdalex',
-    rows: 2,
-    cols: 2,
-  },
-  {
-    id: 10,
-    img: 'https://images.unsplash.com/photo-1567306301408-9b74779a11af',
-    title: 'Tomato basil',
-    author: '@shelleypauls',
-  },
-  {
-    id: 11,
-    img: 'https://images.unsplash.com/photo-1471357674240-e1a485acb3e1',
-    title: 'Sea star',
-    author: '@peterlaster',
-  },
-  {
-    id: 12,
-    img: 'https://images.unsplash.com/photo-1589118949245-7d38baf380d6',
-    title: 'Bike',
-    author: '@southside_customs',
-    cols: 2,
-  },
-];
-
-export default ListBasicComponent;
