+++ src/components/post/CopyUrlComponent.jsx
@@ -6,7 +6,7 @@ const currentUrl = window.location.href;
 const CopyUrlComponent = () => {
 
     return (
-        <CopyToClipboard text={currentUrl} onCopy={()=> alert('주소 카피캣 =^,,^=')}>
+        <CopyToClipboard text={currentUrl} onCopy={() => alert('주소 카피캣 =^,,^=')}>
             <Button>URL 복사</Button>
         </CopyToClipboard>
     )
