--- src/pages/board/AddByEditorPage.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import AddByEditorComponent
-  from "../../components/board/AddByEditorComponent.jsx";
-
-export default function AddByEditorPage() {
-  return (
-      <>
-        <div>
-          보드 글 작성하기
-        </div>
-        <AddByEditorComponent/>
-      </>
-  );
-}
