--- src/pages/board/AddPage.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import AddComponent from "../../components/board/AddComponent.jsx";
-
-function AddPage(props) {
-  return (
-      <>
-        <div>
-          보드 글 작성하기
-        </div>
-        <AddComponent/>
-      </>
-  );
-}
-
-export default AddPage;
