--- src/pages/board/list/ListBasicPage.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import ListBasicComponent
-  from "../../../components/board/list/ListBasicComponent.jsx";
-
-function ListBasicPage() {
-  return (
-      <>
-        <ListBasicComponent/>
-      </>
-  );
-}
-
-export default ListBasicPage;
