+++ src/pages/post/AddByEditorPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import AddByEditorComponent from "../../components/post/AddByEditorComponent.jsx";
+
+export default function AddByEditorPage() {
+    return (
+        <>
+            <div>
+                게시글 작성하기
+            </div>
+            <AddByEditorComponent/>
+        </>
+    );
+}
