+++ src/pages/post/AddPage.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import AddComponent from "../../components/post/AddComponent.jsx";
+
+function AddPage(props) {
+    return (
+        <>
+            <div>
+                게시글 작성하기
+            </div>
+            <AddComponent/>
+        </>
+    );
+}
+
+export default AddPage;
