+++ src/pages/post/ModifyByEditorPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ModifyByEditorComponent from "../../components/board/ModifyByEditorComponent.jsx";
+import ModifyByEditorComponent from "../../components/post/ModifyByEditorComponent.jsx";
 import {useParams} from "react-router-dom";
 
 function ModifyByEditorPage() {
