+++ src/pages/post/ViewerPage.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import ViewerComponent from "../../components/post/ViewerComponent.jsx";
+import {useSelector} from "react-redux";
+
+function ViewerPage(props) {
+
+    const content = useSelector(state => state.postSlice)
+
+    return (
+        <div>
+            <ViewerComponent content={content}/>
+        </div>
+    );
+}
+
+export default ViewerPage;
