+++ src/pages/post/list/ListBasicPage.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import ListBasicComponent from "../../../components/post/list/ListBasicComponent.jsx";
+
+function ListBasicPage() {
+    return (
+        <>
+            <ListBasicComponent/>
+        </>
+    );
+}
+
+export default ListBasicPage;
