+++ src/pages/post/list/ListMyPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ListMyComponent from "../../../components/post/list/ListMyComponent.jsx";
+
+function ListMyPage(props) {
+    return (
+        <>
+            <div>나의 포스트</div>
+            <ListMyComponent/>
+        </>
+    );
+}
+
+export default ListMyPage;
